Start reading addresses from open address table and matching to POIs
This is not perfect right now. It matches on about half of the business addresses. Some of the open addresses have extra spaces/punctuation/abbreviations so we will have to adjust how we are matching if we want greater precision. I don't know if we really care, could honestly be handled with some extra work in the front end.